# Connect refactor to use transactional setState() 