… test/components/connect.spec.js