This is a pull request for https://github.com/reactjs/react-redux/issues/368

Not to be merged yet, as the code is not clean but it's enough to get some feedback, as all the existing tests are passing