@slorber 2016-05-04
@gaearon did you have time to check my PR?

I've tried to run benchmarks of @mweststrate . Not sure to have done them correctly, but I actually see increased performances :) It's not huge, and maybe I did something wrong during the benchmark.

```
# Change 1:
- Before: ~50ms
- After: ~45ms

# Add 1:
- Before: ~110ms
- After: ~103ms
```

Maybe you can run them yourself and confirm?

I've published the benchmark app [before](http://slorber.github.io/connect-refactor/before.html] and [after](http://slorber.github.io/connect-refactor/after.html) so that anyone can take a look