@timdorr 2016-08-14
Superseded by #416